Move protobuf dir
